requirements.txt
python-binance==1.0.19
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
pytest==7.4.0
python-dotenv==1.0.0
aiohttp==3.8.5
schedule==1.2.0
